% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size.R
\name{effect_size_d}
\alias{effect_size_d}
\title{Compute Cohen's d from means and standard deviation}
\usage{
effect_size_d(mean_h1, mean_h0 = 0, sd)
}
\arguments{
\item{mean_h1}{Mean assumed under H1.}

\item{mean_h0}{Mean assumed under H0 or comparison group mean.}

\item{sd}{Common standard deviation.}
}
\value{
Numeric effect size d.
}
\description{
Compute Cohen's d from means and standard deviation
}
