% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size.R
\name{effect_size_f}
\alias{effect_size_f}
\title{Compute Cohen's f from eta-squared}
\usage{
effect_size_f(eta2)
}
\arguments{
\item{eta2}{Proportion of variance explained.}
}
\value{
Numeric effect size f.
}
\description{
Compute Cohen's f from eta-squared
}
