% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size.R
\name{effect_size_f2}
\alias{effect_size_f2}
\title{Compute Cohen's f2 from R-squared}
\usage{
effect_size_f2(r2)
}
\arguments{
\item{r2}{Coefficient of determination.}
}
\value{
Numeric effect size f2.
}
\description{
Compute Cohen's f2 from R-squared
}
