% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size.R
\name{effect_size_f2_increase}
\alias{effect_size_f2_increase}
\title{Compute f2 for an R-squared increase}
\usage{
effect_size_f2_increase(r2_full, r2_reduced)
}
\arguments{
\item{r2_full}{R-squared for the full model.}

\item{r2_reduced}{R-squared for the reduced model.}
}
\value{
Numeric effect size f2.
}
\description{
Compute f2 for an R-squared increase
}
