% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size.R
\name{effect_size_h}
\alias{effect_size_h}
\title{Compute Cohen's h for two proportions}
\usage{
effect_size_h(p1, p2)
}
\arguments{
\item{p1, p2}{Proportions.}
}
\value{
Numeric h.
}
\description{
Compute Cohen's h for two proportions
}
