% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size.R
\name{effect_size_q}
\alias{effect_size_q}
\title{Compute Cohen's q for two correlations}
\usage{
effect_size_q(r1, r2)
}
\arguments{
\item{r1, r2}{Correlations to compare.}
}
\value{
Numeric Fisher-z difference q.
}
\description{
Compute Cohen's q for two correlations
}
