% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size.R
\name{effect_size_w}
\alias{effect_size_w}
\title{Compute chi-square effect size w}
\usage{
effect_size_w(p0, p1)
}
\arguments{
\item{p0}{Probabilities under H0.}

\item{p1}{Probabilities under H1.}
}
\value{
Numeric effect size w.
}
\description{
Compute chi-square effect size w
}
