% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size.R
\name{eta2_from_f}
\alias{eta2_from_f}
\title{Convert Cohen's f to eta-squared}
\usage{
eta2_from_f(f)
}
\arguments{
\item{f}{Cohen's f.}
}
\value{
Numeric eta-squared value.
}
\description{
Convert Cohen's f to eta-squared
}
