% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_result.R
\name{format_result_html}
\alias{format_result_html}
\title{Format a ggpower result as structured HTML for Shiny UI}
\usage{
format_result_html(x)
}
\arguments{
\item{x}{A `ggpower_result` object.}
}
\value{
A `shiny.tag` list for `renderUI`.
}
\description{
Format a ggpower result as structured HTML for Shiny UI
}
