% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{package}
\name{ggpower}
\alias{ggpower}
\alias{ggpower-package}
\title{ggpower: package for power transformations, useful e.g. in ggplot}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{scale_color_sqrt}{

You can add scale_color_sqrt directly to a ggplot2 command.
}

\section{Useful functions}{

power_trans, power_breaks and semi_scientific_formatting can be useful in
many scenarios.
}

