% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_registry.R
\name{ggpower_tests}
\alias{ggpower_tests}
\title{List supported statistical power tests}
\usage{
ggpower_tests(domain = NULL, module = NULL)
}
\arguments{
\item{domain}{Optional character vector to filter by domain (`general`,
`biomarker`, `pharma`).}

\item{module}{Optional character vector to filter by app module (`workspace`,
`biomarker`, `clinical`).}
}
\value{
A data frame describing tests available to `power_compute()`.
}
\description{
List supported statistical power tests
}
