% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{lhist}
\alias{lhist}
\title{Histogram with logged x-axis}
\usage{
lhist(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Useful e.g. for log-normal distributions, etc.. Wraps
ggplot+geom_histogram with all the things I usually get wrong (see details)
Type \code{lhist} to see code for copy-pasting.
}
\details{
Currently, this visualization is the one I find the most useful:
\itemize{
\item x-axis is log-transformed (coord_trans, !not! scale_x - otherwise, adding
theoretical distributions does not work)
\item log-spaced bins, otherwise coord_trans throws error and/or bin widths look
extremely uneven
\item y-axis is sqrt-transformed, otherwise you don't see higher values so well
(play around with the example below and you'll see)
\item add theoretical distributions with
\code{geom_line(aes(logseq(x), dlnorm(x, meanlog, sdlog)))}
}
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 x <- rlnorm(100000, meanlog = c(-6,.4), sdlog = 1)
 xseq <- pmax(1e-5, exp(seq(log(.9*min(x)), log(1.1*max(x)), length.out = 100)))

 ggplot()+
   geom_histogram(data=data.frame(x=x), aes(x, stat(density)), breaks=xseq) +
   geom_line(aes(xseq, .5*dlnorm(xseq, -6, 1)), color="blue") +
   geom_line(aes(xseq, .5*dlnorm(xseq, .4, 1)), color="red") +
   coord_trans(x="log", y="sqrt")
 }
}
}
