% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{logseq}
\alias{logseq}
\title{Log-spaced sequence}
\usage{
logseq(x, length = 100, abs_min = 1e-05)
}
\arguments{
\item{x}{numeric vector}

\item{length}{PARAM_DESCRIPTION, Default: 100}

\item{abs_min}{PARAM_DESCRIPTION, Default: 1e-05}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Generates logarithmically spaced sequence covering the
entire range of \code{x} (plus/minus ten percent) and avoiding
numbers smaller or equal to zero.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
