% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size.R
\name{odds_ratio_from_probs}
\alias{odds_ratio_from_probs}
\title{Compute odds ratio from two probabilities}
\usage{
odds_ratio_from_probs(p0, p1)
}
\arguments{
\item{p0}{Baseline event probability.}

\item{p1}{Event probability under H1.}
}
\value{
Numeric odds ratio.
}
\description{
Compute odds ratio from two probabilities
}
