% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{power_trans}
\alias{power_trans}
\title{FUNCTION_TITLE}
\usage{
power_trans(power)
}
\arguments{
\item{power}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[scales]{trans_new}}
}
