% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{scale_x_power}
\alias{scale_x_power}
\alias{scale_y_power}
\title{Power transformations for ggplot2's axis.}
\usage{
scale_x_power(power = 1/2, ...)

scale_y_power(power = 1/2, ...)
}
\arguments{
\item{power}{A value with \code{1/n}, where \code{n} must be an integer number. Default: 1/2}

\item{...}{Other parameters passed to \code{\link[ggplot2]{scale_x_continuous}}.}
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\section{Functions}{
\itemize{
\item \code{scale_x_power}: same for y-axis.

\item \code{scale_y_power}: same for y-axis.
}}

\examples{
\dontrun{
if(interactive()){
 p <- tibble(a=rgamma(1000, shape = c(.1,3, 16), rate=1)) \%>\% ggplot(aes(a))+
 geom_histogram(bins=50)
 p + scale_x_sqrt()      #  no ticks below 10
 p + scale_x_power(1/2)  #  reasonable ticks (powers of two with human-readable rounding)
 p + scale_x_power(1/4)  #  other powers than .5 are possible
 }
}
}
\seealso{
\code{\link[ggplot2]{scale_continuous}}
 \code{\link[ggpower]{power_trans}},\code{\link[ggpower]{semi_scientific_formatting}}
}
