% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{semi_scientific_formatting}
\alias{semi_scientific_formatting}
\title{FUNCTION_TITLE}
\usage{
semi_scientific_formatting(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[scales]{scientific_format}}
}
