% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{power_breaks}
\alias{power_breaks}
\title{FUNCTION_TITLE}
\usage{
power_breaks(lims, power, n_breaks = 100)
}
\arguments{
\item{lims}{PARAM_DESCRIPTION}

\item{power}{PARAM_DESCRIPTION}

\item{n_breaks}{PARAM_DESCRIPTION, Default: 5}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
