% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{semi_scientific_formatting}
\alias{semi_scientific_formatting}
\title{FUNCTION_TITLE}
\usage{
semi_scientific_formatting(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
# without, with mild and with strong transformation:
 ggplot(data=tibble(x=rgamma(1000, c(.5, 5, 50))), aes(x))+geom_histogram(bins=50)
 ggplot(data=tibble(x=rgamma(1000, c(.5, 5, 50))), aes(x))+geom_histogram(bins=50) + scale_x_continuous(trans=power_trans(1/2))
 ggplot(data=tibble(x=rgamma(1000, c(.5, 5, 50))), aes(x))+geom_histogram(bins=50) + scale_x_continuous(trans=power_trans(1/10))

# last plot, but with nicer x-axis labels:
 ggplot(data=tibble(x=rgamma(1000, c(.5, 5, 50))), aes(x))+geom_histogram(bins=50) + scale_x_continuous(trans=power_trans(1/2), labels = semi_scientific_formatting)
 }
}
}
\seealso{
\code{\link[scales]{scientific_format}}
}
